refactor: move TestsTable to its own folder and make it reusable
- move TestsTable to /src/components/TestsTable
- make TestTable receive the columns as props so we can reuse it
with different columns
