+++ dashboard/src/components/TestsTable/DefaultTestsColumns.tsx
@@ -1,0 +1,139 @@
+import type { ColumnDef } from '@tanstack/react-table';
+
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/Tooltip';
+
+import type { TIndividualTest, TPathTests } from '@/types/general';
+
+import { GroupedTestStatus } from '@/components/Status/Status';
+
+import { TableHeader } from '@/components/Table/TableHeader';
+
+import { ChevronRightAnimate } from '@/components/AnimatedIcons/Chevron';
+
+import TooltipHardware from '@/components/TestsTable/TooltipHardware';
+import { TooltipDateTime } from '@/components/TooltipDateTime';
+
+export const defaultColumns: ColumnDef<TPathTests>[] = [
+  {
+    accessorKey: 'path_group',
+    header: ({ column }): JSX.Element =>
+      TableHeader({
+        column: column,
+        sortable: true,
+        intlKey: 'testDetails.path',
+        intlDefaultMessage: 'Path',
+      }),
+  },
+  {
+    accessorKey: 'pass_tests',
+    header: ({ column }): JSX.Element =>
+      TableHeader({
+        column: column,
+        sortable: true,
+        intlKey: 'testDetails.status',
+        intlDefaultMessage: 'Status',
+        tooltipId: 'bootsTab.statusTooltip',
+      }),
+    cell: ({ row }): JSX.Element => {
+      return (
+        <GroupedTestStatus
+          pass={row.original.pass_tests}
+          done={row.original.done_tests}
+          miss={row.original.miss_tests}
+          fail={row.original.fail_tests}
+          skip={row.original.skip_tests}
+          error={row.original.error_tests}
+        />
+      );
+    },
+  },
+  {
+    id: 'chevron',
+    cell: (): JSX.Element => <ChevronRightAnimate />,
+  },
+];
+
+export const defaultInnerColumns: ColumnDef<TIndividualTest>[] = [
+  {
+    accessorKey: 'path',
+    header: ({ column }): JSX.Element =>
+      TableHeader({
+        column: column,
+        sortable: true,
+        intlKey: 'testDetails.path',
+        intlDefaultMessage: 'Path',
+      }),
+    cell: ({ row }): JSX.Element => {
+      return (
+        <Tooltip>
+          <TooltipTrigger>
+            <div className="max-w-80 overflow-clip text-ellipsis text-nowrap">
+              {row.getValue('path')}
+            </div>
+          </TooltipTrigger>
+          <TooltipContent>{row.getValue('path')}</TooltipContent>
+        </Tooltip>
+      );
+    },
+  },
+  {
+    accessorKey: 'status',
+    header: ({ column }): JSX.Element =>
+      TableHeader({
+        column: column,
+        sortable: true,
+        intlKey: 'testDetails.status',
+        intlDefaultMessage: 'Status',
+      }),
+  },
+  {
+    accessorKey: 'start_time',
+    header: ({ column }): JSX.Element =>
+      TableHeader({
+        column: column,
+        sortable: true,
+        intlKey: 'global.date',
+        intlDefaultMessage: 'Date',
+      }),
+    cell: ({ row }): JSX.Element => (
+      <div className="text-nowrap">
+        <TooltipDateTime
+          dateTime={row.getValue('start_time')}
+          showLabelTime={true}
+        />
+      </div>
+    ),
+  },
+  {
+    accessorKey: 'duration',
+    header: ({ column }): JSX.Element =>
+      TableHeader({
+        column: column,
+        sortable: true,
+        intlKey: 'testDetails.duration',
+        intlDefaultMessage: 'Duration',
+      }),
+    cell: ({ row }): string =>
+      row.getValue('duration') ? row.getValue('duration') : '-',
+  },
+  {
+    id: 'hardware',
+    accessorKey: 'hardware',
+    header: ({ column }): JSX.Element =>
+      TableHeader({
+        column: column,
+        sortable: true,
+        intlKey: 'treeDetails.hardware',
+        intlDefaultMessage: 'Hardware',
+      }),
+    cell: ({ row }): JSX.Element => (
+      <div className="text-nowrap">
+        <TooltipHardware hardwares={row.getValue('hardware')} />
+      </div>
+    ),
+  },
+  {
+    id: 'chevron',
+    cell: (): JSX.Element => <ChevronRightAnimate />,
+  },
+];
